// app/_types/types.ts
export type NodeType = "readFile" | "summarize" | "email" | "report" | "note";
// ...
export type Action =
  | { type: "ready"; payload: FlowNode | undefined }
  | { type: "setLabel"; payload: string }
  | { type: "setDescription"; payload: string }
  | { type: "setReportFormat"; payload: string }
  | { type: "setEmail"; payload: string }
  | { type: "setFile"; payload: { fileName: string; filePath: string } }
  | { type: "removeFile" }
  | { type: "reset"; payload: FlowNode | undefined };